requirements.txt
label-studio-sdk
python-dotenv
Pillow
pandas
numpy